WebSockets/tests/submissions/Opera/opening-handshake/005.html

<!doctype html>
<title>WebSockets: proper first line</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../constants.js></script>
<div id=log></div>
<script>
async_test(function(t) {
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/ws/simple_handshake');
  ws.onmessage = t.step_func(function() {assert_unreached()});
  ws.onopen = function(e) {
    
    ws.onclose = function(e) {
      
      assert_equals(e.wasClean, true);
      ws.onclose = t.step_func(function() {assert_unreached()});
      
      setTimeout(t.step_func(function() {t.done();}), 50)
    }
    ws.close();
  }
});
</script>
